+++ src/text-context.c
@@ -315,13 +315,13 @@ sp_text_context_root_handler (SPEventContext *ec, GdkEvent *event)
 			NRPointF dtp;
 			/* Button 1, set X & Y & new item */
 			sp_selection_empty (SP_DT_SELECTION (ec->desktop));
-			sp_desktop_w2d_xy_point (ec->desktop, &dtp, event->button.x, event->button.y);
+			sp_desktop_w2d_xy_point (ec->desktop, &dtp, (float) event->button.x, (float) event->button.y);
 			sp_desktop_dt2root_xy_point (ec->desktop, &tc->pdoc, dtp.x, dtp.y);
 			/* Cursor */
 			tc->show = TRUE;
 			tc->phase = 1;
 			sp_canvas_item_show (tc->cursor);
-			sp_desktop_w2d_xy_point (ec->desktop, &dtp, event->button.x, event->button.y);
+			sp_desktop_w2d_xy_point (ec->desktop, &dtp, (float) event->button.x, (float) event->button.y);
 			sp_ctrlline_set_coords (SP_CTRLLINE (tc->cursor), dtp.x, dtp.y, dtp.x + 32, dtp.y);
 			/* Processed */
 			return TRUE;
